Adds computed and regular bps to debugger
